Crate ab_contracts_macros

#[contract_impl] macro will process public methods annotated with following attributes:

#[init] - method that can be called to produce an initial state of the contract, called once during contacts lifetime

#[update] - method that can read and/or modify state and/or slots of the contact, may be called by user transaction directly or by another contract

#[view] - method that can only read blockchain data, can read state or slots of the contract, but can’t modify their contents

Each argument (except self) of these methods has to be annotated with one of the following attributes:

#[env] - environment variable, used to access ephemeral execution environment, call methods on other contracts, etc.

#[slot] - slot corresponding to this contract

#[input] - method input coming from user transaction or invocation from another contract

#[output] - method output

#[result] - a single optional method result as an alternative to returning values from a function directly, useful to reduce stack usage

#[init]

Initializer’s purpose is to produce the initial state of the contract.

Following arguments are supported by this method (must be in this order):

#[env] read-only and read-write

#[slot] read-only and read-write

#[input]

#[output]

#[result]

self argument is not supported in any way in this context since state of the contract is just being created.

#[update]

Generic method into contract that can both update contract’s own state and contents of slots.

Following arguments are supported by this method (must be in this order):

&self or &mut self depending on whether state reads and/or modification are required

#[env] read-only and read-write

#[slot] read-only and read-write

#[input]

#[output]

#[result]

#[view]

Similar to #[update], but can only access read-only view of the state and slots, can be called outside of block context and can only call other #[view] methods.

Following arguments are supported by this method (must be in this order):

&self

#[env] read-only

#[slot] read-only

#[input]

#[output]

#[result]

Attribute Macros

contract_impl

#[contract_impl] macro to derive smart contract implementation, see module description for details.